Implemented how lineage tests and reorganized benchmark module
