+++ datatracer/how_lineage/__init__.py
@@ -1,7 +1,18 @@
 from datatracer.how_lineage.base import HowLineageSolver
 from datatracer.how_lineage.basic import BasicHowLineageSolver
+from datatracer.how_lineage.table_transforms import entries_count,\
+    entries_sum, entries_min, entries_max, entries_std, columns_avg,\
+    columns_diff, columns_sum
 
 __all__ = (
     'HowLineageSolver',
     'BasicHowLineageSolver',
+    'entries_count',
+    'entries_sum',
+    'entries_min',
+    'entries_max',
+    'entries_std',
+    'columns_sum',
+    'columns_diff',
+    'columns_avg'
 )
