+++ datatracer/how_lineage/transformer.py
@@ -72,7 +72,7 @@ class Transformer:
             X.append(parent_table["_tmp_"].fillna(0.0).values)
             columns.append({"source_col": {"table_name": fk['table'], "col_name": fk['field']},
                         "row_map": fk,
-                        "aggregation": "datatracer.how_lineage.count"
+                        "aggregation": "datatracer.how_lineage.entries_count"
                         })
 
         return np.array(X).transpose(), columns
@@ -87,10 +87,10 @@ class Transformer:
                 continue
 
             for op, op_name, op_str in [
-                (lambda x: x.sum(), "SUM", "datatracer.how_lineage.sum"),
-                (lambda x: x.max(), "MAX", "datatracer.how_lineage.max"),
-                (lambda x: x.min(), "MIN", "datatracer.how_lineage.min"),
-                (lambda x: x.std(), "STD", "datatracer.how_lineage.std"),
+                (lambda x: x.sum(), "SUM", "datatracer.how_lineage.entries_sum"),
+                (lambda x: x.max(), "MAX", "datatracer.how_lineage.entries_max"),
+                (lambda x: x.min(), "MIN", "datatracer.how_lineage.entries_min"),
+                (lambda x: x.std(), "STD", "datatracer.how_lineage.entries_std"),
             ]:
                 # Count the number of rows for each key.
                 child_table = self.tables[fk["table"]].copy()
